Alert

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { View, FlatList, Text } from 'react-native';
+import { View, FlatList, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { baseUrl } from '../shared/baseUrl';
@@ -33,11 +33,26 @@ class Favorites extends Component {
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert('Delete Favorite?',
+                            'Are you sure you want to delete the favorite dish ' + item.name + ' ?',
+                            [
+                                {
+                                    text: 'Cancel',
+                                    onPress: () => console.log(item.name + ' Not Deleted'),
+                                    style: ' cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            { cancelable: false })
+                    }
                 }
             ];
-    
-            return(
+
+            return (
                 <Swipeout right={rightButton} autoClose={true}>
                     <ListItem
                         key={index}
@@ -45,26 +60,26 @@ class Favorites extends Component {
                         subtitle={item.description}
                         hideChevron={true}
                         onPress={() => navigate('Dishdetail', { dishId: item.id })}
-                        leftAvatar={{ source: { uri: baseUrl + item.image }}}
+                        leftAvatar={{ source: { uri: baseUrl + item.image } }}
                     />
                 </Swipeout>
             );
         };
 
         if (this.props.dishes.isLoading) {
-            return(
+            return (
                 <Loading />
             );
         }
         else if (this.props.dishes.errMess) {
-            return(
+            return (
                 <View>
                     <Text>{this.props.dishes.errMess}</Text>
                 </View>
             );
         }
         else {
-            return(
+            return (
                 <FlatList
                     data={this.props.dishes.dishes.filter(dish => this.props.favorites.some(el => el === dish.id))}
                     renderItem={renderMenuItem}
